requirements.txt
streamlit>=1.32.0
langchain-groq>=0.1.3
langchain-community>=0.2.0
langgraph>=0.1.0
requests>=2.31.0
python-dotenv>=1.0.0
duckduckgo-search>=5.3.0
Pillow>=10.0.0
